Add replpad-js

+++ build-assets.bash
@@ -1,14 +1,27 @@
-[ $R3 ] || {
-  echo "Please export R3=/path/to/r3-android-executable"
-  exit 1
-}
+# CONFIG
+
+function error { echo $1; exit 1; }
+
+[ $R3 ] || error "Please export R3=/path/to/r3-android-executable"
+
+i=`which zip`
+[ $i ] || error "Please put zip in PATH"
 
 rm -rf assets
 mkdir assets
+mkdir -p assets-src/system
+
+#
 
 cd assets-src
+
 zip -r0 ../assets/install.zip \
-  system/
+  system/ \
+  replpad-js/
+
 cp install.sh ../assets
 cp $R3 ../assets
-cd
+
+cd .. # base dir
+
+# vim: set et sw=2:
